move StdOutAdapterSettings to .../adapters/settings.py
