+++ screenpy/narration/adapters/__init__.py
@@ -3,13 +3,14 @@ Default Adapters for the Narrator's microphone.
 """
 
 
-from .stdout_adapter import StdOutAdapter, StdOutManager, settings
+from .settings import config
+from .stdout_adapter import StdOutAdapter, StdOutManager
 
 LoggingAdapter = StdOutAdapter
 
 __all__ = [
+    "config",
     "LoggingAdapter",
-    "settings",
     "StdOutAdapter",
     "StdOutManager",
 ]
