+++ screenpy/narration/adapters/stdout_adapter.py
@@ -7,14 +7,10 @@ from contextlib import contextmanager
 from functools import wraps
 from typing import Any, Callable, Generator, List, Optional
 
-from screenpy.settings import StdOutAdapterSettings
-
 from ..gravitas import AIRY, EXTREME, HEAVY, LIGHT, NORMAL
+from .settings import config
 
 # pylint: disable=unused-argument
-
-# initialize adapter settings
-settings = StdOutAdapterSettings()
 
 
 class StdOutManager:
@@ -23,8 +19,8 @@ class StdOutManager:
     def __init__(self, logger: Optional[logging.Logger] = None) -> None:
         self.logger = logger or logging.getLogger("screenpy")
         self.depth: List[str] = []
-        self.whitespace = settings.INDENT_SIZE * settings.INDENT_CHAR
-        self.enabled = settings.INDENT_LOGS
+        self.whitespace = config.INDENT_SIZE * config.INDENT_CHAR
+        self.enabled = config.INDENT_LOGS
 
     @contextmanager
     def _indent(self) -> Generator:
